requirements.txt
streamlit
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.7
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.129
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.25.0
numpy==1.26.4
openai==1.51.0
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1

